Resolving check_data function issues #4

+++ app/forms.py
@@ -10,7 +10,7 @@ from django.utils.translation import ugettext_lazy as _
 
 class UserInput(forms.Form):
     city = forms.CharField(label='Miasto', initial="Kraków", required=True)
-    price = forms.FloatField(label='Cena mieszkania', initial="500000", required=True)
+    price = forms.IntegerField(label='Cena mieszkania', initial="500000", required=True)
     size_min = forms.IntegerField(label='Wielkość minimum', initial="40", required=True)
     size_max = forms.IntegerField(label='Wielkość maksimum', initial="60", required=True)
     rooms_min = forms.IntegerField(label='Liczba pokoi minimum', initial="2", required=True)
